Gehalt Kalkulation implementiert

+++ PSM-Library/model/Report.cs
@@ -6,7 +6,7 @@ namespace PSM_Libary.model
     {
         public int Id { get; set; }
         public DateTime Date { get; set; }
-        public double Hours { get; set; }
+        public int Hours { get; set; }
         public int EmployeeId { get; set; }
         public int ActivityId { get; set; }
         public int OrderId { get; set; }
